+++ code/cell.modules/pkg.sys/net/src/web.PeerNetwork/types/index.ts
@@ -1,9 +1,13 @@
 export * from './types';
-export * from './types.status';
+export * from './types.peer';
 export * from './types.uri';
 
-export * from './types.events.peer';
-export * from './types.events.group';
+export * from './types.event.peer';
+export * from './types.event.group';
+export * from './types.event.peer.local';
+export * from './types.event.peer.conn';
+export * from './types.event.peer.data';
+export * from './types.event.peer.remote';
 
 export * from './types.strategy';
 export * from './types.group';
