+++ code/cell.modules/pkg.sys/net/src/web.PeerNetwork/types/types.event.peer.ts
@@ -1,0 +1,7 @@
+import { t } from './common';
+
+export type PeerEvent =
+  | t.PeerLocalEvent
+  | t.PeerConnectionEvent
+  | t.PeerDataEvent
+  | t.PeerRemoteEvent;
